+++ notification/sse.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"github.com/labstack/echo"
 	"github.com/satori/go.uuid"
+	"github.com/traPtitech/traQ/model"
 	"github.com/traPtitech/traQ/notification/events"
 	"net/http"
 )
@@ -48,6 +49,24 @@ func (s *sseStreamer) run() {
 		}
 	}
 }
+
+func GetNotificationStream(c echo.Context) error {
+	userId := uuid.FromStringOrNil(c.Get("user").(*model.User).ID)
+
+	if _, ok := c.Response().Writer.(http.Flusher); !ok {
+		return echo.NewHTTPError(http.StatusNotImplemented, "Server Sent Events is not supported.")
+	}
+
+	//Set headers for SSE
+	c.Response().Header().Set(echo.HeaderContentType, "text/event-stream")
+	c.Response().Header().Set("Cache-Control", "no-cache")
+	c.Response().Header().Set("Connection", "keep-alive")
+	c.Response().WriteHeader(http.StatusOK)
+
+	Stream(userId, c.Response())
+	return nil
+}
+
 
 func Stream(userId uuid.UUID, res *echo.Response) {
 	client := &sseClient{
