--- router/notification.go
@@ -1,26 +1,0 @@
-package router
-
-import (
-	"github.com/labstack/echo"
-	"github.com/satori/go.uuid"
-	"github.com/traPtitech/traQ/model"
-	"github.com/traPtitech/traQ/notification"
-	"net/http"
-)
-
-func GetNotificationStream(c echo.Context) error {
-	userId := uuid.FromStringOrNil(c.Get("user").(*model.User).ID)
-
-	if _, ok := c.Response().Writer.(http.Flusher); !ok {
-		return echo.NewHTTPError(http.StatusNotImplemented, "Server Sent Events is not supported.")
-	}
-
-	//Set headers for SSE
-	c.Response().Header().Set(echo.HeaderContentType, "text/event-stream")
-	c.Response().Header().Set("Cache-Control", "no-cache")
-	c.Response().Header().Set("Connection", "keep-alive")
-	c.Response().WriteHeader(http.StatusOK)
-
-	notification.Stream(userId, c.Response())
-	return nil
-}
